+++ src/electron/preload.ts
@@ -6,6 +6,7 @@ import { iTunesImport } from './import_itunes'
 window.addon = addon
 window.isDev = is.dev
 window.isMac = is.mac
+window.isWindows = is.windows
 window.iTunesImport = iTunesImport
 
 window.joinPaths = (...args) => {